UPDATE: swagger; watchlist model and service

+++ src/components/watchlist/watchlist.service.js
@@ -1,25 +1,26 @@
 const WatchlistItemModel = require("./watchlist.model");
 const config = require("../../../config");
 const { isValidId } = require("../../shared/db/db.utils");
+const { ValidationError } = require("../../shared/errors");
 
 async function getAll() {
   const watchListItems = await WatchlistItemModel.find();
-  return watchListItems;
+  return watchListItems.map((item) => basicDetails(item));
 }
 
 async function getOneById(id) {
-  const watchListItems = await WatchlistItemModel.findOne({ id });
-  return watchListItems;
+  const watchListItem = await WatchlistItemModel.findOne({ _id: id });
+  return basicDetails(watchListItem);
 }
 
 async function getAllByUserId(userId) {
   const watchListItems = await WatchlistItemModel.find({ userId });
-  return watchListItems;
+  return watchListItems.map((item) => basicDetails(item));
 }
 
 async function getOneByUserById(id, userId) {
-  const watchListItems = await WatchlistItemModel.findOne({ id, userId });
-  return watchListItems;
+  const watchListItem = await WatchlistItemModel.findOne({ _id: id, userId });
+  return basicDetails(watchListItem);
 }
 
 async function create(itemDetails) {
@@ -28,14 +29,17 @@ async function create(itemDetails) {
   // save item
   await newItemDTO.save();
 
-  return newItemDTO;
+  return basicDetails(newItemDTO);
 }
 
 async function update(id, itemDetails) {
-  const itemDTO = await getOneByUserById(id, itemDetails.userId);
+  const itemDTO = await WatchlistItemModel.findOne({
+    _id: id,
+    userId: itemDetails.userId,
+  });
 
   if (!itemDTO) {
-    throw new Error("Watchlist item does not exist");
+    throw new ValidationError("Watchlist item does not exist");
   }
 
   if (itemDetails.isWatched) {
@@ -47,12 +51,41 @@ async function update(id, itemDetails) {
   itemDTO.dateUpdated = Date.now();
   await itemDTO.save();
 
-  return itemDTO;
+  return basicDetails(itemDTO);
 }
 
 async function remove(id, userId) {
-  const itemDTO = await getOneByUserById(id, userId);
+  const itemDTO = await WatchlistItemModel.findOne({ _id: id, userId });
+  if (!itemDTO) throw new ValidationError("Watchlist item does not exist");
   await itemDTO.deleteOne();
+}
+
+function basicDetails(watchListItem) {
+  const {
+    id,
+    title,
+    director,
+    genre,
+    rating,
+    isWatched,
+    dateAdded,
+    dateUpdated,
+    dateWatched,
+    userId,
+  } = watchListItem;
+
+  return {
+    id,
+    title,
+    director,
+    genre,
+    rating,
+    isWatched,
+    dateAdded,
+    dateUpdated,
+    dateWatched,
+    userId,
+  };
 }
 
 module.exports = {
